Reworks chart updating to support series and pies.
The following new (series aware) functions have been introduced:

  * `line_append_point(chart_id, series, x, y)` which appends a new point
    to the given series of the named line chart.
  * `line_add_point(chart_id, series, x, y)` which shifts a new point
    into the given series of the named line chart.

  * `bar_update(chart_id, series, x, y)` which updates a series in a bar
    chart. Note that bar charts use numbers to indicate categories (x). If
    the given x already exists, it will be updated, otherwise it will be
    added to the chart.

  * `pie_update(chart_id, series, name, y)` which updates a slice in a pier
    chart. Note that pie charts use strings to indicate the slices. If the
    given name already exists, the slice will be updated, otherwise it will
    be added to the chart.

The old `add_point` and `append_point` functions are unaware of series
and can only be used to update the first series of the line chart and
the bar chart.

The new charts.html and charts.eca give a quick overview of the new functions.
(Be warned: the line chart gets big fast! Try this out with a small tweet
collection)